+++ packages/api/lib/controllers/statisticsController.js
@@ -1,13 +1,16 @@
 'use strict';
 
 const { Box, Measurement } = require('@sensebox/opensensemap-api-models'),
-  { UnprocessableEntityError } = require('restify-errors'),
+  { UnprocessableEntityError, BadRequestError } = require('restify-errors'),
   idwTransformer = require('../transformers/idwTransformer'),
-  { addCache } = require('../helpers/apiUtils'),
+  { addCache, createDownloadFilename } = require('../helpers/apiUtils'),
   { retrieveParameters, validateFromToTimeParams } = require('../helpers/userParamHelpers'),
   area = require('@turf/area'),
   millify = require('millify'),
-  handleError = require('../helpers/errorHandler');
+  handleError = require('../helpers/errorHandler'),
+  ms = require('ms'),
+  csvstringify = require('csv-stringify'),
+  DescriptiveStatisticsTransformer = require('../transformers/descriptiveStatisticsTransformer');
 
 /**
  * @api {get} /stats Get some statistics about the database
@@ -122,6 +125,117 @@ const idwHandler = function (req, res, next) {
 
 };
 
+/**
+ * @api {get} /statistics/descriptive Compute basic descriptive statistics over specified time windows
+ * @apiDescription Todo
+ * @apiGroup Descriptive statistics
+ * @apiName descriptiveStatistics
+ * @apiParam {String} phenomenon the name of the phenomenon you want to download the data for.
+ * @apiUse BBoxParam
+ * @apiUse ExposureFilterParam
+ * @apiParam {RFC3339Date} from-date Beginning date of measurement data (default: 2 days ago from now)
+ * @apiParam {RFC3339Date} to-date End date of measurement data (default: now)
+ * @apiParam {String=arithmeticMean,geometricMean,harmonicMean,max,median,min,mode,rootMeanSquare,standardDeviation,sum,variance} operation Statistical operation to execute
+ * @apiParam {String} window Time window to apply. Either a number in Milliseconds or a [`zeit/ms`](https://npmjs.com/ms)-parseable string. At least 1 minute
+ * @apiParam {Boolean=true,false} [download=true] Set the `content-disposition` header to force browsers to download instead of displaying.
+ */
+const descriptiveStatisticsHandler = function descriptiveStatisticsHandler (req, res, next) {
+  const { boxId, bbox, exposure, delimiter, columns, phenomenon, operation, download } = req._userParams;
+  let { fromDate, toDate, window } = req._userParams;
+
+  window = ms(window);
+  if (!window || window < 60000) {
+    return next(new BadRequestError('Invalid window length. Smallest window size is 1 minute.'));
+  }
+
+  // compute start and end times in milliseconds
+  //
+  fromDate = fromDate.valueOf() - (fromDate.valueOf() % window);
+  // always overshoot one averageWindow..
+  toDate = toDate.valueOf() - (toDate.valueOf() % window) + window;
+
+  const windows = [new Date(fromDate)];
+  // compute all possible windows
+  for (let i = 1; windows[i - 1].getTime() !== toDate; i = i + 1) {
+    windows.push(new Date(fromDate + window * i));
+  }
+
+  fromDate = new Date(fromDate);
+  toDate = new Date(toDate);
+
+  if (boxId && bbox) {
+    return next(new BadRequestError('please specify only boxId or bbox'));
+  } else if (!boxId && !bbox) {
+    return next(new BadRequestError('please specify either boxId or bbox'));
+  }
+
+  const opts = {
+    query: {
+      'sensors.title': phenomenon,
+      'exposure': { '$not': /mobile/ }
+    },
+    bbox,
+    from: fromDate,
+    to: toDate,
+    // add sensorId, value and createdAt to query columns
+    columns: ['createdAt', 'sensorId', 'value', ...columns],
+    order: { sensor_id: 1, createdAt: 1 },
+    transformations: { parseValues: true },
+  };
+
+  if (boxId) {
+    opts.query['_id'] = { '$in': boxId };
+  }
+
+  // exposure parameter
+  if (exposure) {
+    // remove mobile boxes
+    if (exposure.includes('mobile')) {
+      exposure.splice(exposure.indexOf('mobile', 1));
+    }
+    opts.query['exposure'] = { '$in': exposure };
+  }
+
+  Box.findMeasurementsOfBoxesStream(opts)
+    .then(function (cursor) {
+      res.header('Content-Type', 'text/csv');
+      if (download === 'true') {
+        res.header('Content-Disposition', `attachment; filename=${createDownloadFilename(req.date(), operation, [phenomenon, ...columns], 'csv')}`);
+      }
+
+      // construct a new array with columns for csv stringify in correct order
+      // (sensorId, <user specified columns>, <averageWindows>)
+      // Start with sensorId. It should always be the the first column
+      // append the wanted columns and the windows
+      const csvColumns = ['sensorId', ...columns, ...windows];
+
+      cursor
+        .on('error', function (err) {
+          return handleError(err, next);
+        })
+        .pipe(new DescriptiveStatisticsTransformer({
+          operation,
+          windows
+        }))
+        .on('error', function (err) {
+          return handleError(err, next);
+        })
+        .pipe(csvstringify({
+          columns: csvColumns, delimiter, header: 1, formatters: {
+            date: d => d.toISOString()
+          }
+        }))
+        .on('error', function (err) {
+          return handleError(err, next);
+        })
+        .pipe(res);
+    })
+    .catch(function (err) {
+      handleError(err, next);
+    });
+
+};
+
 module.exports = {
   getStatistics: [
     retrieveParameters([
@@ -145,5 +259,23 @@ module.exports = {
     ]),
     validateFromToTimeParams,
     idwHandler
+  ],
+  descriptiveStatisticsHandler: [
+    retrieveParameters([
+      { name: 'boxId', aliases: ['senseboxid', 'senseboxids', 'boxid', 'boxids'], dataType: ['id'] },
+      { name: 'phenomenon', required: true },
+      { predef: 'delimiter' },
+      { name: 'exposure', allowedValues: Box.BOX_VALID_EXPOSURES, dataType: ['String'] },
+      { name: 'columns', dataType: ['String'], defaultValue: [], allowedValues: ['boxId', 'boxName', 'exposure', 'height', 'lat', 'lon', 'phenomenon', 'sensorType', 'unit'] },
+      { predef: 'bbox' },
+      { predef: 'toDate', required: true },
+      { predef: 'fromDate', required: true },
+      { name: 'window', required: true },
+      { name: 'operation', required: true, allowedValues: ['arithmeticMean', 'geometricMean', 'harmonicMean', 'max', 'median', 'min', 'mode', 'rootMeanSquare', 'standardDeviation', 'sum', 'variance'] },
+      { name: 'includeMobile', defaultValue: 'false', allowedValues: ['true', 'false'] },
+      { name: 'download', defaultValue: 'true', allowedValues: ['true', 'false'] }
+    ]),
+    validateFromToTimeParams,
+    descriptiveStatisticsHandler
   ]
 };