+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## Unreleased
 - Initial Release after splitting api and models
 - add transformer for classifying the state of boxes
+- Add descriptive statistics measurements transformer
+- Add descriptive statistics route